Support benchmarks with a very large number of subscores
For instance, stylebench has 429 subscores. The URL we fetch for data from
Treeherder was over 8000 characters long which is about twice as long as allowed.

This change makes parallel requests of up to 100 signatures IDs at a time.
The lenght of the URLs are about 2000 characters about half of what Treeherder allows.